+++ app/models/ingest_request.rb
@@ -5,10 +5,10 @@ class IngestRequest < ApplicationRecord
   scope :pending_ingest, -> { where('ingested_at IS NULL') }
 
   validates_presence_of :uuid
-  validate :not_already_pending_validation, on: :create
+  validate :not_already_pending_validation, if: :new_record?
 
   attr_accessor :test_mode
-  after_create :send_to_fedora, on: :create
+  after_create :send_to_fedora
 
   private