+++ app/models/mms_client.rb
@@ -3,7 +3,7 @@
 require 'http'
 require 'nokogiri'
 
-class MMSClient
+class MmsClient
   def initialize(options = {})
     @url = options[:mms_url]
     @basic_username = options[:user_name]
@@ -335,7 +335,8 @@ class MMSClient
   # This DRYs up the pattern of making a request and throwing an exception for bad responses
   #   - A 410 response means the capture is part of a deleted Item.
   def make_request_for(export_type, uuid, params = {})
-    response = authed_request.get(export_url_for(export_type, uuid), params: params)
+    export_url = export_url_for(export_type, uuid)
+    response = authed_request.get(export_url, params: params)
 
     # Return nil for things that have moved permanently. This will allow the update to go through in a limited fashion.
     if response.code == 410